Extract utils and types into @storybook/core-common and @storybook/types
