+++ code/lib/cli/src/helpers.ts
@@ -8,13 +8,13 @@ import stripJsonComments from 'strip-json-comments';
 import findUp from 'find-up';
 import invariant from 'tiny-invariant';
 import { getCliDir, getRendererDir } from './dirs';
-import type {
-  JsPackageManager,
-  PackageJson,
-  PackageJsonWithDepsAndDevDeps,
+import {
+  type JsPackageManager,
+  type PackageJson,
+  type PackageJsonWithDepsAndDevDeps,
+  frameworkToRenderer as CoreFrameworkToRenderer,
 } from '@storybook/core-common';
-import type { SupportedFrameworks } from '@storybook/types';
-import type { SupportedRenderers } from './project_types';
+import type { SupportedFrameworks, SupportedRenderers } from '@storybook/types';
 import { CoreBuilder } from './project_types';
 import { SupportedLanguage } from './project_types';
 import { versions as storybookMonorepoPackages } from '@storybook/core-common';
@@ -132,40 +132,10 @@ type CopyTemplateFilesOptions = {
   destination?: string;
 };
 
-export const frameworkToRenderer: Record<
-  SupportedFrameworks | SupportedRenderers,
-  SupportedRenderers | 'vue'
-> = {
-  // frameworks
-  angular: 'angular',
-  ember: 'ember',
-  'html-vite': 'html',
-  'html-webpack5': 'html',
-  nextjs: 'react',
-  'preact-vite': 'preact',
-  'preact-webpack5': 'preact',
-  qwik: 'qwik',
-  'react-vite': 'react',
-  'react-webpack5': 'react',
-  'server-webpack5': 'server',
-  solid: 'solid',
-  'svelte-vite': 'svelte',
-  'svelte-webpack5': 'svelte',
-  sveltekit: 'svelte',
-  'vue3-vite': 'vue3',
-  'vue3-webpack5': 'vue3',
-  'web-components-vite': 'web-components',
-  'web-components-webpack5': 'web-components',
-  // renderers
-  html: 'html',
-  preact: 'preact',
-  'react-native': 'react-native',
-  react: 'react',
-  server: 'server',
-  svelte: 'svelte',
-  vue3: 'vue3',
-  'web-components': 'web-components',
-};
+/**
+ * @deprecated Please use `frameworkToRenderer` from `@storybook/core-common` instead
+ */
+export const frameworkToRenderer = CoreFrameworkToRenderer;
 
 export const frameworkToDefaultBuilder: Record<SupportedFrameworks, CoreBuilder> = {
   angular: CoreBuilder.Webpack5,
