+++ code/lib/types/src/index.ts
@@ -10,3 +10,4 @@ export * from './modules/indexer';
 export * from './modules/composedStory';
 export * from './modules/channelApi';
 export * from './modules/frameworks';
+export * from './modules/renderers';
