+++ code/lib/types/src/modules/renderers.ts
@@ -1,0 +1,14 @@
+// Should match @storybook/<renderer>
+export type SupportedRenderers =
+  | 'react'
+  | 'react-native'
+  | 'vue3'
+  | 'angular'
+  | 'ember'
+  | 'preact'
+  | 'svelte'
+  | 'qwik'
+  | 'html'
+  | 'web-components'
+  | 'server'
+  | 'solid';
